slab-slub-slob-convert-slab_flags_t-to-32-bit-fix
fix printk warning

mm/slub.c:3660: warning: format '%lx' expects type 'long unsigned int', but argument 7 has type 'slab_flags_t'

Cc: Alexey Dobriyan <adobriyan@gmail.com>

+++ mm/slub.c
@@ -3657,7 +3657,7 @@ error:
 	if (flags & SLAB_PANIC)
 		panic("Cannot create slab %s size=%lu realsize=%u order=%u offset=%u flags=%lx\n",
 		      s->name, (unsigned long)s->size, s->size,
-		      oo_order(s->oo), s->offset, flags);
+		      oo_order(s->oo), s->offset, (unsigned long)flags);
 	return -EINVAL;
 }
